+++ src/App.vue
@@ -1,4 +1,9 @@
 <script lang="ts" setup>
+import Help from './components/Help.vue'
+import Markmap from './components/Markmap.vue'
+import Pen from './components/Pen.vue'
+import ControlPanel from './components/ControlPanel.vue'
+
 import { useMarkmap } from '@/stores/markmap'
 const markmapStore = useMarkmap()
 </script>
@@ -8,6 +13,7 @@ const markmapStore = useMarkmap()
     class="relative"
     :class="!markmapStore.manual && markmapStore.theme === 'dark' ? 'dark' : ''"
   >
+    <ControlPanel />
     <Markmap />
     <Help />
     <Pen />