+++ tailwind.config.js
@@ -1,11 +1,41 @@
+const animate = require('tailwindcss-animate')
+
+/** @type {import('tailwindcss').Config} */
 module.exports = {
-  content: ['index.html', 'src/**/*.vue', 'src/stores/*.ts'],
-  darkMode: 'class',
+  darkMode: ['class'],
+  content: [
+    './index.html',
+    './src/stores/*.ts',
+    './pages/**/*.{ts,tsx,vue}',
+    './components/**/*.{ts,tsx,vue}',
+    './app/**/*.{ts,tsx,vue}',
+    './src/**/*.{ts,tsx,vue}',
+  ],
+  prefix: '',
   theme: {
-    extend: {},
+    container: {
+      center: true,
+      padding: '2rem',
+      screens: {
+        '2xl': '1400px',
+      },
+    },
+    extend: {
+      keyframes: {
+        'accordion-down': {
+          from: { height: 0 },
+          to: { height: 'var(--radix-accordion-content-height)' },
+        },
+        'accordion-up': {
+          from: { height: 'var(--radix-accordion-content-height)' },
+          to: { height: 0 },
+        },
+      },
+      animation: {
+        'accordion-down': 'accordion-down 0.2s ease-out',
+        'accordion-up': 'accordion-up 0.2s ease-out',
+      },
+    },
   },
-  variants: {
-    extend: {},
-  },
-  plugins: [],
+  plugins: [animate],
 }